fix: the detail for purchase bill

+++ pkg/model/bill.go
@@ -64,7 +64,7 @@ type BillProductResponse struct {
 	TotalBeforeVAT string `json:"total_before_vat"`
 	TotalVAT       string `json:"total_vat"`
 	Total          string `json:"total"`
-	Type           int16
+	Type           int8
 }
 
 type AddProductToBillParams struct {
